+++ spec/rgame/engine/components/velocity_spec.rb
@@ -21,8 +21,8 @@ RSpec.describe RGame::Engine::Components::Velocity do
   end
 
   it_behaves_like 'a mover' do
-    def build_mover(blocked_by:, heading: [1, 0])
-      described_class.new(vx: 60.0 * heading[0], vy: 60.0 * heading[1], blocked_by: blocked_by)
+    def build_mover(blocked_by:, pushes: [], heading: [1, 0])
+      described_class.new(vx: 60.0 * heading[0], vy: 60.0 * heading[1], blocked_by:, pushes:)
     end
   end